Extract residual_graph_of from Dinic into max_flow
The counting sort that builds the residual graph, caches the capacity of each
arc's pair in its padding and merges parallel arcs sat inside
Dinic::from_edge_list. A second solver wants the same graph built the same way,
so it moves to max_flow beside the trait and the arc types it works on.

No behaviour changes. Dinic keeps its own node count, since it sizes its
scratch from it, and calls the shared builder for the rest.
